homework_4/login_microservice/login_microservice/requirements.txt
Django>=3.2,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
gunicorn
whitenoise
pandas